Improvements to socket finding, added license, cleanup
The host side is now a bit smarter about finding the socket file. Both
the host side and the client side can now be passed a socket file name
by command arguments.

Added GPLv2 as the license. This may change if I end up making a
graphical option for the client side (which would likely use PyQt, thus
requiring me to use GPL3).

As of this commit, kate has been removed as the editor in favour of
smarter default editor detection. Set EDITOR to contain 'kate -b' to
recreate the previous behaviour.

The portion of the argument parser which selects the shell has been
improved so it works as its own argument group. Mutually exclusive
groups don't have all the features of standard argument groups, so the
shell arguments are in a mutex group which is the only element in an
argument group. This may go away eventually, but not until Python
versions with this limitation are unpopular.