Moved packet handling into its own shared module.
Much of the code for handling specific packets was previously inside of
sshed_client.SocketRequestHandler. I moved it out into a separate
PacketHandler class (in a new shared module) and made
SocketRequestHandler inherit from PacketHandler.

It was then simple to replace the packet handling code in sshed.py with
an instance of PacketHandler.

Since diff based editing is defined in a packetised manner, this will
allow diff based editing to be added fairly simply.